+++ src/DocChecks.jl
@@ -8,13 +8,11 @@ import ..Documenter:
     Documenter,
     Documents,
     Utilities,
-    Utilities.Markdown2
+    Utilities.Markdown2,
+    Utilities.@docerror
 
 using DocStringExtensions
 import Markdown
-
-using Logging
-loglevel(doc) = doc.user.strict ? Logging.Error : Logging.Warn
 
 # Missing docstrings.
 # -------------------
@@ -65,8 +63,7 @@ function missingdocs(doc::Documents.Document)
         println(b, """\n
             These are docstrings in the checked modules (configured with the modules keyword)
             that are not included in @docs or @autodocs blocks.""")
-        push!(doc.internal.errors, :missing_docs)
-        @logmsg loglevel(doc) String(take!(b))
+        @docerror(doc, :missing_docs, String(take!(b)))
     end
 end
 
@@ -128,18 +125,15 @@ function footnotes(doc::Documents.Document)
         for (id, (ids, bodies)) in orphans
             # Multiple footnote bodies.
             if bodies > 1
-                push!(doc.internal.errors, :footnote)
-                @logmsg loglevel(doc) "footnote '$id' has $bodies bodies in $(Utilities.locrepr(page.source))."
+                @docerror(doc, :footnote, "footnote '$id' has $bodies bodies in $(Utilities.locrepr(page.source)).")
             end
             # No footnote references for an id.
             if ids === 0
-                push!(doc.internal.errors, :footnote)
-                @logmsg loglevel(doc) "unused footnote named '$id' in $(Utilities.locrepr(page.source))."
+                @docerror(doc, :footnote, "unused footnote named '$id' in $(Utilities.locrepr(page.source)).")
             end
             # No footnote bodies for an id.
             if bodies === 0
-                push!(doc.internal.errors, :footnote)
-                @logmsg loglevel(doc) "no footnotes found for '$id' in $(Utilities.locrepr(page.source))."
+                @docerror(doc, :footnote, "no footnotes found for '$id' in $(Utilities.locrepr(page.source)).")
             end
         end
     end
@@ -181,8 +175,7 @@ function linkcheck(doc::Documents.Document)
                 end
             end
         else
-            push!(doc.internal.errors, :linkcheck)
-            @logmsg loglevel(doc) "linkcheck requires `curl`."
+            @docerror(doc, :linkcheck, "linkcheck requires `curl`.")
         end
     end
     return nothing
@@ -208,8 +201,7 @@ function linkcheck(link::Markdown.Link, doc::Documents.Document; method::Symbol=
             # interpolating into backticks escapes spaces so constructing a Cmd is necessary
             result = read(cmd, String)
         catch err
-            push!(doc.internal.errors, :linkcheck)
-            @logmsg loglevel(doc) "$cmd failed:" exception = err
+            @docerror(doc, :linkcheck, "$cmd failed:", exception = err)
             return false
         end
         STATUS_REGEX = r"^(\d+) (\w+)://(?:\S+) (\S+)?$"m
@@ -239,12 +231,10 @@ function linkcheck(link::Markdown.Link, doc::Documents.Document; method::Symbol=
                 @debug "linkcheck '$(link.url)' status: $(status), retrying without `-I`"
                 return linkcheck(link, doc; method=:GET)
             else
-                push!(doc.internal.errors, :linkcheck)
-                @logmsg loglevel(doc) "linkcheck '$(link.url)' status: $(status)."
+                @docerror(doc, :linkcheck, "linkcheck '$(link.url)' status: $(status).")
             end
         else
-            push!(doc.internal.errors, :linkcheck)
-            @logmsg loglevel(doc) "invalid result returned by $cmd:" result
+            @docerror(doc, :linkcheck, "invalid result returned by $cmd:", result)
         end
     end
     return false