+++ src/Expanders.jl
@@ -16,7 +16,7 @@ import .Documents:
     EvalNode,
     MetaNode
 
-import .Utilities: Selectors
+import .Utilities: Selectors, @docerror
 
 import Markdown, REPL
 import Base64: stringmime
@@ -256,8 +256,8 @@ function Selectors.runner(::Type{MetaBlocks}, x, page, doc)
             try
                 meta[ex.args[1]] = Core.eval(Main, ex.args[2])
             catch err
-                push!(doc.internal.errors, :meta_block)
-                @warn("""
+                @docerror(doc, :meta_block,
+                    """
                     failed to evaluate `$(strip(str))` in `@meta` block in $(Utilities.locrepr(page.source, lines))
                     ```$(x.language)
                     $(x.code)
@@ -283,8 +283,8 @@ function Selectors.runner(::Type{DocsBlocks}, x, page, doc)
         binding = try
             Documenter.DocSystem.binding(curmod, ex)
         catch err
-            push!(doc.internal.errors, :docs_block)
-            @warn("""
+            @docerror(doc, :docs_block,
+                """
                 unable to get the binding for '$(strip(str))' in `@docs` block in $(Utilities.locrepr(page.source, lines)) from expression '$(repr(ex))' in module $(curmod)
                 ```$(x.language)
                 $(x.code)
@@ -296,8 +296,8 @@ function Selectors.runner(::Type{DocsBlocks}, x, page, doc)
         end
         # Undefined `Bindings` get discarded.
         if !Documenter.DocSystem.iskeyword(binding) && !Documenter.DocSystem.defined(binding)
-            push!(doc.internal.errors, :docs_block)
-            @warn("""
+            @docerror(doc, :docs_block,
+                """
                 undefined binding '$(binding)' in `@docs` block in $(Utilities.locrepr(page.source, lines))
                 ```$(x.language)
                 $(x.code)
@@ -311,8 +311,8 @@ function Selectors.runner(::Type{DocsBlocks}, x, page, doc)
         object = Utilities.Object(binding, typesig)
         # We can't include the same object more than once in a document.
         if haskey(doc.internal.objects, object)
-            push!(doc.internal.errors, :docs_block)
-            @warn("""
+            @docerror(doc, :docs_block,
+                """
                 duplicate docs found for '$(strip(str))' in `@docs` block in $(Utilities.locrepr(page.source, lines))
                 ```$(x.language)
                 $(x.code)
@@ -332,8 +332,8 @@ function Selectors.runner(::Type{DocsBlocks}, x, page, doc)
 
         # Check that we aren't printing an empty docs list. Skip block when empty.
         if isempty(docs)
-            push!(doc.internal.errors, :docs_block)
-            @warn("""
+            @docerror(doc, :docs_block,
+                """
                 no docs found for '$(strip(str))' in `@docs` block in $(Utilities.locrepr(page.source, lines))
                 ```$(x.language)
                 $(x.code)
@@ -383,8 +383,8 @@ function Selectors.runner(::Type{AutoDocsBlocks}, x, page, doc)
                     fields[ex.args[1]] = Core.eval(curmod, ex.args[2])
                 end
             catch err
-                push!(doc.internal.errors, :autodocs_block)
-                @warn("""
+                @docerror(doc, :autodocs_block,
+                    """
                     failed to evaluate `$(strip(str))` in `@autodocs` block in $(Utilities.locrepr(page.source, lines))
                     ```$(x.language)
                     $(x.code)
@@ -451,8 +451,8 @@ function Selectors.runner(::Type{AutoDocsBlocks}, x, page, doc)
         nodes = DocsNode[]
         for (mod, path, category, object, isexported, docstr) in results
             if haskey(doc.internal.objects, object)
-                push!(doc.internal.errors, :autodocs_block)
-                @warn("""
+                @docerror(doc, :autodocs_block,
+                    """
                     duplicate docs found for '$(object.binding)' in $(Utilities.locrepr(page.source, lines))
                     ```$(x.language)
                     $(x.code)
@@ -475,8 +475,8 @@ function Selectors.runner(::Type{AutoDocsBlocks}, x, page, doc)
         end
         page.mapping[x] = DocsNodes(nodes)
     else
-        push!(doc.internal.errors, :autodocs_block)
-        @warn("""
+        @docerror(doc, :autodocs_block,
+            """
             '@autodocs' missing 'Modules = ...' in $(Utilities.locrepr(page.source, lines))
             ```$(x.language)
             $(x.code)
@@ -502,8 +502,8 @@ function Selectors.runner(::Type{EvalBlocks}, x, page, doc)
             try
                 result = Core.eval(sandbox, ex)
             catch err
-                push!(doc.internal.errors, :eval_block)
-                @warn("""
+                @docerror(doc, :eval_block,
+                    """
                     failed to evaluate `@eval` block in $(Utilities.locrepr(page.source))
                     ```$(x.language)
                     $(x.code)
@@ -587,13 +587,13 @@ function Selectors.runner(::Type{ExampleBlocks}, x, page, doc)
             result = c.value
             print(buffer, c.output)
             if c.error
-                push!(doc.internal.errors, :example_block)
-                @warn("""
+                @docerror(doc, :example_block,
+                    """
                     failed to run `@example` block in $(Utilities.locrepr(page.source, lines))
                     ```$(x.language)
                     $(x.code)
                     ```
-                    """, c.value)
+                    """, value = c.value)
                 page.mapping[x] = x
                 return
             end
@@ -656,7 +656,6 @@ function Selectors.runner(::Type{REPLBlocks}, x, page, doc)
     @debug "Evaluating @repl block:\n$(x.code)"
     for (ex, str) in Utilities.parseblock(x.code, doc, page; keywords = false,
                                           linenumbernode = linenumbernode)
-        buffer = IOBuffer()
         input  = droplines(str)
         if VERSION >= v"1.5.0-DEV.178"
             # Use the REPL softscope for REPLBlocks,
@@ -670,11 +669,12 @@ function Selectors.runner(::Type{REPLBlocks}, x, page, doc)
         end
         Core.eval(mod, Expr(:global, Expr(:(=), :ans, QuoteNode(c.value))))
         result = c.value
+        buf = IOContext(IOBuffer(), :color=>ansicolor)
         output = if !c.error
             hide = REPL.ends_with_semicolon(input)
-            Documenter.DocTests.result_to_string(buffer, hide ? nothing : c.value)
+            Documenter.DocTests.result_to_string(buf, hide ? nothing : c.value)
         else
-            Documenter.DocTests.error_to_string(buffer, c.value, [])
+            Documenter.DocTests.error_to_string(buf, c.value, [])
         end
         if !isempty(input)
             push!(multicodeblock, Markdown.Code("julia-repl", prepend_prompt(input)))
@@ -714,8 +714,8 @@ function Selectors.runner(::Type{SetupBlocks}, x, page, doc)
         end
         Markdown.MD([])
     catch err
-        push!(doc.internal.errors, :setup_block)
-        @warn("""
+        @docerror(doc, :setup_block,
+            """
             failed to run `@setup` block in $(Utilities.locrepr(page.source))
             ```$(x.language)
             $(x.code)