+++ src/Writers/HTMLWriter.jl
@@ -78,13 +78,17 @@ struct HTMLAsset
     class :: Symbol
     uri :: String
     islocal :: Bool
+    attributes::Dict{Symbol, String}
 
-    function HTMLAsset(class::Symbol, uri::String, islocal::Bool)
+    function HTMLAsset(class::Symbol, uri::String, islocal::Bool, attributes::Dict{Symbol, String}=Dict{Symbol,String}())
         if !islocal && match(r"^https?://", uri) === nothing
             error("Remote asset URL must start with http:// or https://")
         end
+        if islocal && isabspath(uri)
+            @error("Local asset should not have an absolute URI: $uri")
+        end
         class in [:ico, :css, :js] || error("Unrecognised asset class $class for `$(uri)`")
-        new(class, uri, islocal)
+        new(class, uri, islocal, attributes)
     end
 end
 
@@ -123,7 +127,7 @@ Documenter.HTML(assets = [
 ])
 ```
 """
-function asset(uri; class = nothing, islocal=false)
+function asset(uri; class = nothing, islocal=false, attributes=Dict{Symbol,String}())
     if class === nothing
         class = assetclass(uri)
         (class === nothing) && error("""
@@ -131,7 +135,7 @@ function asset(uri; class = nothing, islocal=false)
         It can be set explicitly with the `class` keyword argument.
         """)
     end
-    HTMLAsset(class, uri, islocal)
+    HTMLAsset(class, uri, islocal, attributes)
 end
 
 function assetclass(uri)
@@ -312,12 +316,13 @@ This allows search engines to know which version to send their users to. [See
 wikipedia for more information](https://en.wikipedia.org/wiki/Canonical_link_element).
 Default is `nothing`, in which case no canonical link is set.
 
-**`analytics`** can be used specify the Google Analytics tracking ID.
-
 **`assets`** can be used to include additional assets (JS, CSS, ICO etc. files). See below
 for more information.
 
-**`ansicolor`** can be used to enable/disable colored output from `@repl` and `@example` blocks globally.
+**`analytics`** can be used specify the Google Analytics tracking ID.
+
+**`collapselevel`** controls the navigation level visible in the sidebar. Defaults to `2`.
+To show fewer levels by default, set `collapselevel = 1`.
 
 **`sidebar_sitename`** determines whether the site name is shown in the sidebar or not.
 Setting it to `false` can be useful when the logo already contains the name of the package.
@@ -338,13 +343,15 @@ passing an instance of [`KaTeX`](@ref), [`MathJax2`](@ref), or
 [`MathJax3`](@ref) objects, respectively. The rendering engine can further be customized by
 passing options to the [`KaTeX`](@ref) or [`MathJax2`](@ref)/[`MathJax3`](@ref) constructors.
 
-**`lang`** specifies the [`lang` attribute](https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/lang)
-of the top-level `<html>` element, declaring the language of the generated pages. The default
-value is `"en"`.
-
 **`footer`** can be a valid single-line markdown `String` or `nothing` and is displayed below
 the page navigation. Defaults to `"Powered by [Documenter.jl](https://github.com/JuliaDocs/Documenter.jl)
 and the [Julia Programming Language](https://julialang.org/)."`.
+
+**`ansicolor`** can be used to enable/disable colored output from `@repl` and `@example` blocks globally.
+
+**`lang`** specifies the [`lang` attribute](https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/lang)
+of the top-level `<html>` element, declaring the language of the generated pages. The default
+value is `"en"`.
 
 **`warn_outdated`** inserts a warning if the current page is not the newest version of the
 documentation.
@@ -608,7 +615,7 @@ module RD
         ))
     end
     function mathengine!(r::RequireJS, engine::MathJax3)
-        url = isempty(engine.url) ? "https://cdnjs.cloudflare.com/ajax/libs/mathjax/3.1.4/es5/tex-svg.js" : engine.url
+        url = isempty(engine.url) ? "https://cdnjs.cloudflare.com/ajax/libs/mathjax/3.2.0/es5/tex-svg.js" : engine.url
         push!(r, Snippet([], [],
             """
             window.MathJax = $(json_jsescape(engine.config, 2));
@@ -959,15 +966,16 @@ function render_head(ctx, navnode)
 end
 
 function asset_links(src::AbstractString, assets::Vector{HTMLAsset})
+    isabspath(src) && @error("Absolute path '$src' passed to asset_links")
     @tags link script
     links = DOM.Node[]
     for asset in assets
         class = asset.class
         url = asset.islocal ? relhref(src, asset.uri) : asset.uri
         node =
-            class == :ico ? link[:href  => url, :rel => "icon", :type => "image/x-icon"] :
-            class == :css ? link[:href  => url, :rel => "stylesheet", :type => "text/css"] :
-            class == :js  ? script[:src => url] : continue # Skip non-js/css files.
+            class == :ico ? link[:href  => url, :rel => "icon", :type => "image/x-icon", pairs(asset.attributes)...] :
+            class == :css ? link[:href  => url, :rel => "stylesheet", :type => "text/css", pairs(asset.attributes)...] :
+            class == :js  ? script[:src => url, pairs(asset.attributes)...] : continue # Skip non-js/css files.
         push!(links, node)
     end
     return links
@@ -1412,6 +1420,7 @@ function generate_redirect_file(redirectfile::AbstractString, entries)
     comment = "<!--This file is automatically generated by Documenter.jl-->"
 
     isfile(redirectfile) && !startswith(read(redirectfile, String), comment) && return
+    isempty(entries) && return
 
     open(redirectfile, "w") do buf
         println(buf, comment)
@@ -1484,7 +1493,7 @@ end
 
 function domify(lb::ListBuilder)
     @tags ul li
-    ul(map(e -> isa(e, ListBuilder) ? domify(e) : li(e), lb.es))
+    ul(map(e -> isa(e, ListBuilder) ? li[".no-marker"](domify(e)) : li(e), lb.es))
 end
 
 function domify(ctx, navnode, contents::Documents.ContentsNode)